+++ R/reconstruct_pattern.R
@@ -6,8 +6,9 @@
 #' @param n_random Number of randomizations.
 #' @param e_threshold Minimum energy to stop reconstruction.
 #' @param max_runs Maximum number of iterations of e_threshold is not reached.
+#' @param no_change Reconstrucction will stop if energy does not decrease for this number of iterations.
 #' @param fitting It true, the pattern reconstruction starts with a fitting of a Thomas process.
-#' @param comp_fast Should summary functions be estimated in an computational fast way.
+#' @param comp_fast If pattern contains more points than threshold, summary functions are estimated in a computational fast way.
 #' @param return_input The original input data is returned as last list entry
 #' @param simplify If n_random = 1 and return_input = FALSE only pattern will be returned.
 #' @param verbose Print progress report.
@@ -19,10 +20,15 @@
 #' algorithm starts with a random reconstructed pattern, shifts a point to a new location and
 #' keeps the change only, if the deviation between the observed and the reconstructed
 #' pattern decreases. The pair correlation function and the nearest neighbour
-#' distance function are used to describe the patterns. For large patterns
-#' `comp_fast = TRUE` decreases the computational demand because no edge
-#' correction is used and the pair correlation function is estimated based on Ripley's
-#' K-function. For more information see \code{\link{estimate_pcf_fast}}.
+#' distance function are used to describe the patterns.
+#'
+#' For large patterns (\code{n > comp_fast}) the pair correlation function can be estimated
+#' from Ripley's K-function without edge correction. This decreases the computational
+#' time. For more information see \code{\link{estimate_pcf_fast}}.
+#'
+#' The reconstruction can be stopped automatically if for n steps the energy does not
+#' decrease. The number of steps can be controlled by \code{no_change} and is set to
+#' \code{no_change = Inf} as default to never stop automatically.
 #'
 #' @seealso
 #' \code{\link{calculate_energy}} \cr
@@ -46,9 +52,13 @@
 #' in ecology. Boca Raton: Chapman and Hall/CRC Press.
 #'
 #' @export
-reconstruct_pattern <- function(pattern, n_random = 19,
-                                e_threshold = 0.01, max_runs = 10000,
-                                fitting = FALSE, comp_fast = FALSE,
+reconstruct_pattern <- function(pattern,
+                                n_random = 1,
+                                e_threshold = 0.01,
+                                max_runs = 1000,
+                                no_change = Inf,
+                                fitting = FALSE,
+                                comp_fast = 1000,
                                 return_input = TRUE,
                                 simplify = FALSE,
                                 verbose = TRUE,
@@ -58,6 +68,24 @@ reconstruct_pattern <- function(pattern, n_random = 19,
   if(!n_random >= 1) {
     stop("n_random must be >= 1.", call. = FALSE)
   }
+
+  # check if number of points exceed comp_fast limit
+  if(pattern$n > comp_fast) {
+
+    # Print message that summary functions will be computed fast
+    if(verbose) {
+      message("> Using fast compuation of summary functions.")
+    }
+
+    comp_fast <- TRUE
+  }
+
+  else {
+    comp_fast <- FALSE
+  }
+
+  # counter if energy changed
+  energy_counter <- 0
 
   pattern <- spatstat::unmark(pattern) # only spatial points
 
@@ -91,7 +119,7 @@ reconstruct_pattern <- function(pattern, n_random = 19,
         difference <- simulated$n - pattern$n
 
         # id of points to remove
-        remove_points <- sample(seq_len(pattern$n), size = difference)
+        remove_points <- sample(seq_len(simulated$n), size = difference)
 
         # remove points
         simulated <- simulated[-remove_points]
@@ -152,7 +180,7 @@ reconstruct_pattern <- function(pattern, n_random = 19,
     }
 
     # energy before reconstruction
-    e0 <- mean(abs(gest_observed[[3]] - gest_simulated[[3]]), na.rm = TRUE) +
+    energy <- mean(abs(gest_observed[[3]] - gest_simulated[[3]]), na.rm = TRUE) +
       mean(abs(pcf_observed[[3]] - pcf_simulated[[3]]), na.rm = TRUE)
 
     # random ids of pattern
@@ -164,8 +192,8 @@ reconstruct_pattern <- function(pattern, n_random = 19,
                                       nsim = 1, drop = TRUE,
                                       win = pattern$window)
 
-    # pattern reconstruction algorithm (optimaztion of e0) - not longer than max_runs
-    for(i in seq_len(max_runs)){
+    # pattern reconstruction algorithm (optimaztion of energy) - not longer than max_runs
+    for(i in seq_len(max_runs)) {
 
       relocated <- simulated # data for relocation
 
@@ -199,11 +227,13 @@ reconstruct_pattern <- function(pattern, n_random = 19,
         mean(abs(pcf_observed[[3]] - pcf_relocated[[3]]), na.rm = TRUE)
 
       # lower energy after relocation
-      if(e_relocated < e0){
+      if(e_relocated < energy){
 
         simulated <- relocated # keep relocated pattern
 
-        e0 <- e_relocated # keep e_relocated as e0
+        energy <- e_relocated # keep e_relocated as energy
+
+        energy_counter <- 0
 
         # plot observed vs reconstructed
         if(plot) {
@@ -223,15 +253,21 @@ reconstruct_pattern <- function(pattern, n_random = 19,
         }
       }
 
-      # print progress
-      if(verbose) {
-        cat(paste0("\rProgress: n_random: ", current_pattern, "/", n_random,
-                   " || max_runs: ", i, "/", max_runs,
-                   " || e0 = ", round(e0, 5)))
+      # increase counter no change
+      else {
+        energy_counter <- energy_counter + 1
       }
 
-      # exit loop if e threshold is reached
-      if(e0 <= e_threshold){
+      # print progress
+      if(verbose) {
+        message("\r> Progress: n_random: ", current_pattern, "/", n_random,
+                " || max_runs: ", i, "/", max_runs,
+                " || energy = ", round(energy, 5),
+                appendLF = FALSE)
+      }
+
+      # exit loop if e threshold or no_change counter max is reached
+      if(energy <= e_threshold || energy_counter > no_change){
         break
       }
     }
@@ -242,8 +278,8 @@ reconstruct_pattern <- function(pattern, n_random = 19,
   # add input pattern to randomizations
   if(return_input){
 
-    if(verbose & simplify){
-      cat("\n")
+    if(simplify){
+      message("\n")
       warning("'simplify = TRUE' not possible for 'return_input = TRUE'.", call. = FALSE)
     }
 
@@ -256,8 +292,8 @@ reconstruct_pattern <- function(pattern, n_random = 19,
 
     if(simplify) {
 
-      if(verbose & n_random > 1) {
-        cat("\n")
+      if(n_random > 1) {
+        message("\n")
         warning("'simplify = TRUE' not possible for 'n_random > 1'.", call. = FALSE)
       }
 
@@ -269,6 +305,11 @@ reconstruct_pattern <- function(pattern, n_random = 19,
     else{
       names(result) <- paste0("randomized_", seq_len(n_random)) # set names
     }
+  }
+
+  # write result in new line if progress was printed
+  if(verbose) {
+    message("\r")
   }
 
   return(result)